feat(36): Onboarding and Framing - MCP public-launch quickstart, prerequisites, honest command-tool framing
Phase 36 of v1.5 MCP Public-Launch Hardening (ONBRD-01..03).

- docs/mcp-quickstart.md: public sign-up -> generate key -> paste JSON ->
  open globe tab walkthrough; two capability tiers (read vs command);
  troubleshooting incl. engine_unreachable / no_active_plugins (ONBRD-01).
- ConnectAgentHelper: prerequisites callout (signed-in, cloud edition,
  open-tab requirement for command tools); extracted AGENT_PROMPT and
  CopyField into separate modules (file 352 -> 201 lines) (ONBRD-02).
- Every command tool description now begins with a shared
  SESSION_REQUIRED_PREAMBLE stating the live-open-tab requirement
  unmissably (pan_globe, focus_entity, toggle_layer, set_timeline,
  set_filter, clear_filter, fly_to) (ONBRD-03).
- New src/lib/mcp/toolDescriptionFragments.ts dedups the 7 identical
  preambles into one constant.

Gate: tsc clean, 1004 tests pass, build green.
