+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "worldwideview",
-  "version": "2.48.8",
+  "version": "2.48.9",
   "private": true,
   "type": "module",
   "packageManager": "pnpm@9.15.4",
